+++ dist_node/compile/khepri_peep.js
@@ -85,6 +85,11 @@ addPeephole("CurryExpression", (function(node) {
 }), (function(node) {
     return node.base;
 }));
+addPeephole("CurryExpression", (function(node) {
+    return (node.base.type === "CurryExpression");
+}), (function(node) {
+    return ast_expression.CurryExpression.create(null, node.base.base, concat(node.base.args, node.args));
+}));
 addPeephole("BinaryExpression", (function(node) {
     return ((node.operator === "|>") && ((((node.right.type === "CurryExpression") || (node.right.type ===
         "BinaryOperatorExpression")) || (node.right.type === "UnaryOperatorExpression")) || (node.right