+++ index.js
@@ -1,5 +1,4 @@
 var requireDirectory = require('require-directory');
+var path = require('path');
 
-module.exports = requireDirectory(module, './dist_node');
-
-
+module.exports = requireDirectory(module, path.join(__dirname, 'dist_node'));